Move main logic, add css modules, change icons
