--- packages/client/src/app/types/add.d.ts
@@ -1,4 +1,0 @@
-declare global {
-  export type BarsPointsType = { level: number; hp: number };
-}
-export {}
