+++ packages/client/src/pages/MainPage/types/index.ts
@@ -1,0 +1,6 @@
+export interface BarsPointsType {
+  level: number;
+  hp: number;
+}
+
+export type HandleUpdateBarsPointsType = ({ level, hp }: BarsPointsType) => void;
