+++ packages/client/src/pages/MainPage/ui/MainContainer/index.ts
