+++ packages/client/src/pages/MainPage/ui/MainContainer/ui/ActionBlock/index.ts
