+++ packages/client/src/pages/MainPage/ui/MainContainer/ui/MainContainer.module.scss
