+++ packages/client/src/pages/MainPage/ui/MainFooter/index.ts
