+++ packages/client/src/pages/MainPage/ui/MainFooter/ui/MainFooter.tsx
@@ -1,0 +1,12 @@
+import LeaderboardSvg from '@/shared/assets/images/leaderboard.svg';
+import PreferencesSvg from '@/shared/assets/images/preferences.svg';
+import css from './MainFooter.module.scss';
+
+export const MainFooter = () => {
+  return (
+    <footer className={css.main__footer}>
+      <LeaderboardSvg />
+      <PreferencesSvg />
+    </footer>
+  );
+};
