+++ packages/client/src/pages/MainPage/ui/MainHeader/index.ts
