+++ packages/client/src/pages/MainPage/ui/MainHeader/ui/MainHeader.module.scss
