+++ packages/client/src/pages/MainPage/ui/MainHeader/ui/MainHeader.tsx
@@ -1,5 +1,6 @@
 import { LevelOrHpBar } from '@/shared/ui/LevelOrHpBar';
-import './MainHeader.scss';
+import { BarsPointsType } from '../../../types';
+import css from './MainHeader.module.scss';
 
 type MainHeaderProps = {
   barsPoints: BarsPointsType;
@@ -8,9 +9,9 @@ type MainHeaderProps = {
 export const MainHeader = (props: MainHeaderProps) => {
   const { barsPoints, currentLevel } = props;
   return (
-    <header className='main__header'>
+    <header className={css.main__header}>
       <LevelOrHpBar theme='level' text={currentLevel} stat={barsPoints.level} />
-      <h1 className='main__header__title'>OSKAR</h1>
+      <h1 className={css.main__header__title}>OSKAR</h1>
       <LevelOrHpBar theme='hp' text='HP' stat={barsPoints.hp} />
     </header>
   );
