+++ packages/client/src/pages/MainPage/ui/MainPage.module.scss
@@ -1,0 +1,12 @@
+.main {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-between;
+
+  height: 100vh;
+
+  & > * {
+    user-select: none;
+  }
+}
