+++ packages/client/src/pages/MainPage/ui/MainPage.tsx
@@ -1,15 +1,16 @@
 import { useState } from 'react';
+import { MainContainer } from './MainContainer';
+import { MainFooter } from './MainFooter';
+import { MainHeader } from './MainHeader';
 
-import { MainFooter, MainHeader, MainContainer } from '@/widgets';
-
-import './MainPage.scss';
+import css from './MainPage.module.scss';
 
 export const MainPage = () => {
   const [currentLevel, setCurrentLevel] = useState(1);
   const [barsPoints, setBarsPoints] = useState({ level: 50, hp: 90 });
 
   return (
-    <main className='main'>
+    <main className={css.main}>
       <MainHeader barsPoints={barsPoints} currentLevel={currentLevel} />
       <MainContainer
         barsPoints={barsPoints}
