--- packages/client/src/widgets/index.ts
@@ -1,4 +1,0 @@
-export { MainContainer } from './MainContainer';
-export { ActionBlock } from './ActionBlock';
-export { MainHeader } from './MainHeader';
-export { MainFooter } from './MainFooter';
